adding features to select page and cacheline sizes
Not all architectures have the same page sizes or cacheline sizes.
This commit adds architecture specific features that enable the
selection of differenet page and cacheline sizes. One prominent example
is aarch64 that supports three different granules and thus has features
to select the page size for the relevant granule.

Signed-off-by: Reto Achermann <achreto@gmail.com>